+++ libs/designer/src/lib/ui/mcp/parameters/EditOperation.tsx
@@ -4,15 +4,11 @@ import {
   Field,
   Divider,
   Card,
-  Label,
-  Button,
   Accordion,
   AccordionItem,
   AccordionHeader,
   AccordionPanel,
-  mergeClasses,
 } from '@fluentui/react-components';
-import { Dismiss16Regular } from '@fluentui/react-icons';
 import type { RootState, AppDispatch } from '../../../core/state/mcp/store';
 import { useSelector, useDispatch } from 'react-redux';
 import { useCallback, useMemo } from 'react';
@@ -27,8 +23,7 @@ import {
   type UpdateParametersPayload,
 } from '../../../core/state/operation/operationMetadataSlice';
 import { getGroupIdFromParameterId } from '../../../core/utils/parameters/helper';
-import { ParameterEditor } from './ParameterEditor';
-import constants from '../../../common/constants';
+import { ParameterField } from './parameterfield';
 
 interface EditOperationProps {
   description: string;
@@ -384,70 +379,6 @@ export const EditOperation = ({ description, handleDescriptionInputChange, onPar
           <div className={styles.emptyParametersCard}>
             <Text className={styles.emptyParametersText}>{INTL_TEXT.noParametersMessage}</Text>
           </div>
-        )}
-      </div>
-    </div>
-  );
-};
-
-const ParameterField = ({
-  operationId,
-  groupId,
-  parameter,
-  isConditional,
-  onParameterVisibilityUpdate,
-  handleParameterValueChange,
-  handleRemoveConditionalParameter,
-}: {
-  operationId: string;
-  groupId: string;
-  parameter: ParameterInfo;
-  isConditional?: boolean;
-  onParameterVisibilityUpdate: () => void;
-  handleParameterValueChange: (parameterId: string, newValue: ValueSegment[]) => void;
-  handleRemoveConditionalParameter: (parameterId: string) => void;
-}) => {
-  const intl = useIntl();
-  const styles = useEditOperationStyles();
-
-  const removeParamText = intl.formatMessage({
-    id: '5E66mK',
-    defaultMessage: 'Remove parameter',
-    description: 'Tooltip for remove parameter button',
-  });
-
-  const isLargeParameter = useMemo(() => {
-    const editor = parameter.editor?.toLowerCase();
-    return (
-      editor === constants.EDITOR.ARRAY ||
-      editor === constants.EDITOR.DICTIONARY ||
-      editor === constants.EDITOR.HTML ||
-      editor === constants.EDITOR.TABLE
-    );
-  }, [parameter.editor]);
-
-  return (
-    <div className={styles.parameterField}>
-      <Label className={styles.parameterLabel} required={parameter.required} title={parameter.label}>
-        {parameter.label}
-      </Label>
-      <div className={mergeClasses(styles.parameterValueSection, isLargeParameter && styles.largeParameterSection)}>
-        <ParameterEditor
-          operationId={operationId}
-          groupId={groupId}
-          parameter={parameter}
-          onParameterVisibilityUpdate={onParameterVisibilityUpdate}
-          handleParameterValueChange={handleParameterValueChange}
-        />
-        {isConditional && (
-          <Button
-            appearance="subtle"
-            size="small"
-            icon={<Dismiss16Regular />}
-            onClick={() => handleRemoveConditionalParameter(parameter.id)}
-            title={removeParamText}
-            className={styles.removeParameterButton}
-          />
         )}
       </div>
     </div>
